refactor: make renderer module
